requirements.txt
asgiref==3.8.1
DateTime==5.5
Django==5.1.3
numpy==2.1.3
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
zope.interface==7.1.1
